pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oneshot"
version = "0.1.0"
description = "Autonomous task completion with auditor validation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
maintainers = [
    {name = "Your Name", email = "your.email@example.com"},
]
keywords = ["claude", "ai", "automation", "task-completion"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "anyio>=4.0.0",
    "python-statemachine>=2.0.0",
]

[project.optional-dependencies]
ui = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "rich>=13.0.0",
    "websockets>=11.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/oneshot"
Documentation = "https://github.com/yourusername/oneshot#readme"
Repository = "https://github.com/yourusername/oneshot.git"
Issues = "https://github.com/yourusername/oneshot/issues"
Changelog = "https://github.com/yourusername/oneshot/blob/main/CHANGELOG.md"

[project.scripts]
oneshot = "cli.oneshot_cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--disable-warnings",
]